refactor(Profile): Move profile folder to root path
