+++ src/components/Menu/index.tsx
@@ -35,11 +35,6 @@ const Menu = (props) => {
     return footerLinks(t)
   }, [t])
 
-  const subLinks =
-    activeMenuItem?.hideSubNav || activeSubMenuItem?.hideSubNav || pathname.includes('/profile')
-      ? []
-      : activeMenuItem?.items
-
   return (
     <UikitMenu
       linkComponent={(linkProps) => {
@@ -60,7 +55,7 @@ const Menu = (props) => {
       setLang={setLanguage}
       cakePriceUsd={cakePriceUsd.toNumber()}
       links={menuItems}
-      subLinks={subLinks}
+      subLinks={activeMenuItem?.hideSubNav || activeSubMenuItem?.hideSubNav ? [] : activeMenuItem?.items}
       footerLinks={getFooterLinks}
       activeItem={activeMenuItem?.href}
       activeSubItem={activeSubMenuItem?.href}
