+++ src/pages/profile/[accountAddress].tsx
@@ -1,10 +1,10 @@
 import { useWeb3React } from '@web3-react/core'
 import { useRouter } from 'next/router'
 import { useProfileForAddress } from 'state/profile/hooks'
-import { NftProfileLayout } from 'views/Nft/market/Profile'
-import SubMenu from 'views/Nft/market/Profile/components/SubMenu'
-import UnconnectedProfileNfts from 'views/Nft/market/Profile/components/UnconnectedProfileNfts'
-import UserNfts from 'views/Nft/market/Profile/components/UserNfts'
+import { NftProfileLayout } from 'views/Profile'
+import SubMenu from 'views/Profile/components/SubMenu'
+import UnconnectedProfileNfts from 'views/Profile/components/UnconnectedProfileNfts'
+import UserNfts from 'views/Profile/components/UserNfts'
 import { useNftsForAddress } from 'views/Nft/market/hooks/useNftsForAddress'
 
 const NftProfilePage = () => {
