+++ src/pages/profile/[accountAddress]/activity.tsx
@@ -1,0 +1,16 @@
+import { NftProfileLayout } from 'views/Profile'
+import ActivityHistory from 'views/Profile/components/ActivityHistory'
+import SubMenu from 'views/Profile/components/SubMenu'
+
+const NftProfileActivityPage = () => {
+  return (
+    <>
+      <SubMenu />
+      <ActivityHistory />
+    </>
+  )
+}
+
+NftProfileActivityPage.Layout = NftProfileLayout
+
+export default NftProfileActivityPage
