+++ src/views/Ifos/components/IfoFoldableCard/IfoPoolCard/IfoCardActions.tsx
@@ -4,7 +4,6 @@ import { useWeb3React } from '@web3-react/core'
 import { NextLinkFromReactRouter } from 'components/NextLink'
 import { Ifo, PoolIds } from 'config/constants/types'
 import { WalletIfoData, PublicIfoData } from 'views/Ifos/types'
-import { nftsBaseUrl } from 'views/Nft/market/constants'
 import ConnectWalletButton from 'components/ConnectWalletButton'
 import ContributeButton from './ContributeButton'
 import ClaimButton from './ClaimButton'
@@ -46,7 +45,7 @@ const IfoCardActions: React.FC<Props> = ({
 
   if (!hasProfile) {
     return (
-      <Button as={NextLinkFromReactRouter} to={`${nftsBaseUrl}/profile/${account.toLowerCase()}`} width="100%">
+      <Button as={NextLinkFromReactRouter} to={`/profile/${account.toLowerCase()}`} width="100%">
         {t('Activate your Profile')}
       </Button>
     )
