+++ src/views/Nft/market/components/ProfileCell.tsx
@@ -3,7 +3,6 @@ import { Box, Flex, Skeleton, Text, ProfileAvatar } from '@pancakeswap/uikit'
 import truncateHash from 'utils/truncateHash'
 import { useProfileForAddress } from 'state/profile/hooks'
 import { NextLinkFromReactRouter } from 'components/NextLink'
-import { nftsBaseUrl } from '../constants'
 
 const StyledFlex = styled(Flex)`
   align-items: center;
@@ -19,7 +18,7 @@ const ProfileCell: React.FC<{ accountAddress: string }> = ({ accountAddress }) =
   const profileName = profile?.username || '-'
 
   return (
-    <NextLinkFromReactRouter to={`${nftsBaseUrl}/profile/${accountAddress}`}>
+    <NextLinkFromReactRouter to={`/profile/${accountAddress}`}>
       <StyledFlex>
         {!isFetching ? (
           <ProfileAvatar
