+++ src/views/Nft/market/components/SearchBar.tsx
@@ -1,12 +1,11 @@
 import AddressInputSelect from 'components/AddressInputSelect'
-import { nftsBaseUrl } from 'views/Nft/market/constants'
 import { useRouter } from 'next/router'
 
 const SearchBar: React.FC = (props) => {
   const router = useRouter()
 
   const handleAddressClick = (value: string) => {
-    router.push(`${nftsBaseUrl}/profile/${value}`)
+    router.push(`/profile/${value}`)
   }
 
   return <AddressInputSelect onAddressClick={handleAddressClick} {...props} />
