+++ src/views/Pools/components/ProfileRequirementWarning.tsx
@@ -34,7 +34,7 @@ export function ProfileRequirementWarning({
                 {t('Create Profile')} »
               </NextLinkFromReactRouter>
             ) : (
-              <NextLinkFromReactRouter style={{ textDecoration: 'underline' }} to="/nft/profile">
+              <NextLinkFromReactRouter style={{ textDecoration: 'underline' }} to="/profile">
                 {t('Go to Profile')} »
               </NextLinkFromReactRouter>
             )}
