+++ src/views/Profile/components/EditProfileModal/StartView.tsx
@@ -7,7 +7,7 @@ import { getPancakeProfileAddress } from 'utils/addressHelpers'
 import { useCake } from 'hooks/useContract'
 import { useGetCakeBalance } from 'hooks/useTokenBalance'
 import { useTranslation } from 'contexts/Localization'
-import useGetProfileCosts from 'views/Nft/market/Profile/hooks/useGetProfileCosts'
+import useGetProfileCosts from 'views/Profile/hooks/useGetProfileCosts'
 import { FetchStatus } from 'config/constants/types'
 import { requiresApproval } from 'utils/requiresApproval'
 import { useProfile } from 'state/profile/hooks'
