+++ src/views/ProfileCreation/index.tsx
@@ -3,7 +3,6 @@ import { useWeb3React } from '@web3-react/core'
 import Page from 'components/Layout/Page'
 import { useProfile } from 'state/profile/hooks'
 import PageLoader from 'components/Loader/PageLoader'
-import { nftsBaseUrl } from 'views/Nft/market/constants'
 import { useRouter } from 'next/router'
 import Header from './Header'
 import ProfileCreationProvider from './contexts/ProfileCreationProvider'
@@ -16,7 +15,7 @@ const ProfileCreation = () => {
 
   useEffect(() => {
     if (account && hasProfile) {
-      router.push(`${nftsBaseUrl}/profile/${account.toLowerCase()}`)
+      router.push(`/profile/${account.toLowerCase()}`)
     }
   }, [account, hasProfile, router])
