+++ src/views/TradingCompetition/components/RegisterModal/ReactivateProfile.tsx
@@ -1,7 +1,6 @@
 import { Heading, Button, Text } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import { CompetitionProps } from 'views/TradingCompetition/types'
-import { nftsBaseUrl } from 'views/Nft/market/constants'
 import { useWeb3React } from '@web3-react/core'
 import { useRouter } from 'next/router'
 
@@ -11,7 +10,7 @@ const ReactivateProfile: React.FC<CompetitionProps> = ({ onDismiss }) => {
   const router = useRouter()
 
   const handleClick = () => {
-    router.push(`${nftsBaseUrl}/profile/${account.toLowerCase()}`)
+    router.push(`/profile/${account.toLowerCase()}`)
     onDismiss()
   }
